requirements.txt
zarr
numpy
Pillow
dask
distributed
tqdm
glymur
numcodecs
